V1.11.35
controller_update.sh: no longer overwrites go-bluetooth-start.sh and rfcommServerConfig.py, now can do controller specific actions.
manual_update.py: no longer goes down the complete list to end up only updating one version.
gocontroll-first-install: wifi was made into a kernel module, install script now deploys this module.
flash-led.js: now exits if no leds are present on the controller.
testcan.js: now exits immediately after the test is done if no leds are present on the controller.
added the go-rollback-update alias for /etc/controller_update/controller_update_rollback.py
controller_update_rollback.py: added safety check for the user to reaffirm their decision.

+++ etc/controller_update/controller_update_rollback.py
@@ -1,11 +1,13 @@
 #!/usr/bin/python3
 import os
 
-with open("/etc/controller_update/backed-up-files.txt", "r") as backup:
-    for line in backup.readlines():
-        try:
-            os.replace(line[:-1], line[:-2])
-        except:
-            print("backup file is missing:" + line)
-os.remove("/etc/controller_update/backed-up-files.txt")
-#
+arg = input("are you sure you want to undo the last update? y/n: ")
+
+if arg is "y" or arg is "Y":
+    with open("/etc/controller_update/backed-up-files.txt", "r") as backup:
+        for line in backup.readlines():
+            try:
+                os.replace(line[:-1], line[:-2])
+            except:
+                print("backup file is missing:" + line)
+    os.remove("/etc/controller_update/backed-up-files.txt")
